DEX-893 Column closed_at in orders table
Main:
 * Column closed_at added into orders table
 * Inserts and updates in single batch optimized (unnecessary updates avoided)

Renamings:
 * HistoryRouter renamed to HistoryRouterActor
 * HistoryMessagesBatchSender renamed to HistoryMessagesBatchSenderActor
 * HistoryRouterSpecification renamed to HistoryRouterActorSpecification
